+++ Example/ReusableDemo/TableViewController.swift
@@ -16,7 +16,7 @@ class TableViewController: UITableViewController {
 
     tableView.registerReusableCell(MySimpleColorCell)
     tableView.registerReusableCell(MyXIBTextCell)
-    tableView.registerReusableCell(MyXIBSwitchCell)
+    tableView.registerReusableCell(MyXIBInfoCell)
     /* No need to register this one, the UIStoryboard already auto-register its cells */
 //    tableView.registerReusableCell(MyStoryBoardIndexPathCell)
   }
@@ -51,11 +51,9 @@ class TableViewController: UITableViewController {
       textCell.fill("{section \(indexPath.section), row \(indexPath.row)}")
       return textCell
     case 2:
-      let boolCell = tableView.dequeueReusableCell(indexPath: indexPath) as MyXIBSwitchCell
-      boolCell.fill("Switch #\(indexPath.row)", value: boolValues[indexPath.row]) {
-        [unowned self] in self.boolValues[indexPath.row] = $0
-      }
-      return boolCell
+      let infoCell = tableView.dequeueReusableCell(indexPath: indexPath) as MyXIBInfoCell
+      infoCell.fill("InfoCell #\(indexPath.row)", info: "Info #\(indexPath.row)", details: "Details #\(indexPath.row)")
+      return infoCell
     case 3:
       let pathCell = tableView.dequeueReusableCell(indexPath: indexPath) as MyStoryBoardIndexPathCell
       pathCell.fill(indexPath)